Scaling and Rotation matrices
Two other matrices added for scaling and rotating sprites. Rotation is
only done on the Z axis for now; I'll worry about rotating on other axes
when I get around to implementing 3D model loading.

Decided to remove the MovableSprite class and just have the GameObject
class deal with everything related to matrices this way when Quads and
Ellips are reworked, they'll be able to move properly as well.