Miscellaneous Updates:
Not much happened. I've been working on this on and off for the past two months trying to fix a stuttering issues that have resulted since switching to shaders.
This is probably a really dumb issue, but since I have limited experience with OpenGL besides this project, I'm running into some problems fixing this. I guess it's time to ask reddit what's going on.